change numberInhabitants to double to allow proper scaling and using statistical data of average households

+++ src/test/groovy/edu/ie3/test/common/ThermalUnitInputTestData.groovy
@@ -39,7 +39,7 @@ class ThermalUnitInputTestData extends SystemParticipantTestData {
   private static final ComparableQuantity<Temperature> UPPER_TEMPERATURE_LIMIT = Quantities.getQuantity(25, StandardUnits.TEMPERATURE)
   private static final ComparableQuantity<Temperature> LOWER_TEMPERATURE_LIMIT = Quantities.getQuantity(15, StandardUnits.TEMPERATURE)
   private static final String HOUSING_TYPE = "house"
-  private static final Integer NUMBER_INHABITANTS = 2
+  private static final double NUMBER_INHABITANTS = 2.0
 
   public static final thermalHouseInput = new ThermalHouseInput(
   thermalUnitUuid,
